Soundy 1.0 release

+++ soundy.js
@@ -1,17 +1,29 @@
-const { Discord, ytdl, Client, token, prefix, fs, os, HelpMenu, InviteMenu, tts, util } = require("./Modules/Variables.js");
-const { validateYouTubeUrl, SaveTTS } = require("./Modules/Functions.js");
+const { Discord, Client, token, prefix, fs, os, HelpMenu, InviteMenu, tts, util, ActivitiesList } = require("./Modules/Variables.js");
+const { SaveTTS } = require("./Modules/Functions.js");
 
-var isReady = true;
+isReady = true;
 
 Client.on('ready', () => {
     console.log("Logged in!");
-    Client.user.setActivity("Type \'soundy help\'", { type: "PLAYING" });
+    var SoundyHelp = true
+    setInterval(() => {
+        if (SoundyHelp == true){
+            Client.user.setActivity("Type \"soundy help\"");
+            SoundyHelp = false
+        }else{
+            const index = Math.floor(Math.random() * ActivitiesList.length + 1)
+            Client.user.setActivity(ActivitiesList[index])
+            SoundyHelp = true
+        }
+    }, 10000)
 });
 
 Client.on('message', message => {
+    const args = message.content.slice(prefix.length).trim().split(' ');
+    const command = args.shift()
     function Play(Wtp) {
         try{
-            var voiceChannel = message.member.voice.channel;
+            voiceChannel = message.member.voice.channel;
             if (voiceChannel != null)
             {
                 isReady = false
@@ -20,13 +32,12 @@ Client.on('message', message => {
                     message.guild.me.voice.setSelfDeaf(true)
                     const dispatcher = connection.play(Wtp);
                     dispatcher.on("finish", end => {
-                        console.log("Finished")
                         voiceChannel.leave()
                         isReady = true;
                     })
 
                     dispatcher.on("error", end => {
-                        console.log("Error")
+                        console.error()
                         voiceChannel.leave()
                         isReady = true;
                     })
@@ -37,12 +48,11 @@ Client.on('message', message => {
         }catch(err){
             console.log("Failed to play an audio")
             console.log(err)
+            Client.users.cache.get("561560432100245520").send("An error ocurred playing an audio btw, lemme give you more info: \n```\n" + err + "\n```");
         }
     }
-    const args = message.content.slice(prefix.length).trim().split(' ');
-    const command = args.shift()
     if (!message.content.startsWith(prefix) || message.author.bot) return;
-    else if (isReady || command.toLowerCase() == "stop") {
+    else if (isReady || command.toLowerCase() == "stop" || command.toLowerCase() == "eval" || command.toLowerCase() == "eval") {
         switch(command.toLowerCase()){
             case "stop":
                 if (message.guild.me.voice.channel != null){
@@ -52,7 +62,8 @@ Client.on('message', message => {
                 message.channel.send("Umm what do you want to stop? xD")
                 break
             case "help":
-                message.channel.send(HelpMenu);
+                message.react("☑")
+                message.author.send(HelpMenu);
                 break
             case "flipacoin":
                 if (Math.floor((Math.random() * 2) + 1) == 1){
@@ -62,7 +73,8 @@ Client.on('message', message => {
                 }
                 break
             case "ping":
-                message.channel.send("The ping is \"" + `${message.createdTimestamp - Date.now()}` + " ms\"");
+                let Ping = message.createdTimestamp - Date.now()
+                message.channel.send("The ping is \"" + `${Ping.toString().slice(1)}` + " ms\"");
                 break
             case "invite":
                 message.channel.send(InviteMenu)
@@ -86,36 +98,32 @@ Client.on('message', message => {
  
                     message.channel.send(evaled, {code:"xl"});
                 } catch (err) {
-                    message.channel.send("\`ERROR\` \`\`\`xl\n${(err)}\n\`\`\`");
+                    message.channel.send("\`ERROR\` \`\`\`xl\n " + err + "\n\`\`\`");
                 }
                 break
             case "tts":
-                message.delete();
                 if (args.length > 0){
                         SaveTTS(Play("TTS.mp3"), message.content.split(' ').splice(2).join(' '))
-                        message.reply(" said \"" + message.content.split(' ').splice(2).join(' ') + "\" in VC")
                 }else
                 message.channel.send("It would be good if you uuuh... Provided text")
+                break
+            case "say":
+                if (args.length > 0){
+                    message.delete()
+                    message.channel.send("Sending message...").then((sentMessage) => sentMessage.edit(message.content.split(' ').splice(2).join(' ')))
+                }else
+                message.channel.send("*says nothing*")
+                break
+            case "snipe":
+                message.channel.send("Here you have", {files: ["./TTS.mp3"]})
                 break
             default:
                 try {
                     if (fs.existsSync('./Audio/' + command.toLowerCase() + ".mp3")) {
-                      //file exists
-                      console.log(("./Audio/" + command.toLowerCase() + ".mp3" + " Exists!"))
                       message.delete();
                       Play('./Audio/' + command.toLowerCase() + ".mp3")
                     }else{
-                        console.log("Unexisting file")
-                        try{
-                            if (validateYouTubeUrl(command) == true){
-                                Play(ytdl(command, {quality: 'highestaudio', filter: "audioonly"}))
-                                message.delete();
-                            }else
-                            message.channel.send("I can't find that! o.O")
-                        }catch(err){
-                            message.channel.send("I can't find that! o.O")
-                            return console.log(err);
-                        }
+                        message.channel.send("I can't find that! o.O")
                     }
                 } catch(err) {
                     console.error(err)
@@ -125,7 +133,15 @@ Client.on('message', message => {
     }
     else if (isReady == false)
     {
-        message.channel.send("Hold on a second! I'm busy ")
+        if (isReady == true){
+            message.channel.send("Hold on! An unknown error occured")
+            Client.users.cache.get("561560432100245520").send("An error occured, check it lazy dumbass https://cp.something.host/services/1509%22%22 ");
+            const { Discord, Client, token, prefix, fs, os, HelpMenu, InviteMenu, tts, util, ActivitiesList } = require("./Modules/Variables.js");
+            const { SaveTTS } = require("./Modules/Functions.js");
+
+            isReady = true;
+        }else
+        message.channel.send("Hold on a second! I'm busy")
     }
 });
